auth files renaming and restructuring
- specAuthenticator is the provider's client authenticator interface
- extracted out apiAuth to be in its own file
- extracted out specAPIKeyAuthenticator
- header and query specific implementations of specAPIKeyAuthenticator live
in their own file
